+++ include/tao/pegtl/buffer_input.hpp
@@ -106,6 +106,12 @@ namespace tao
             return m_current.data[ offset ];
          }
 
+         std::uint8_t peek_uint8( const std::size_t offset = 0 ) const noexcept
+         {
+            return static_cast< std::uint8_t >( peek_char( offset ) );
+         }
+
+         // Compatibility, remove with 3.0
          std::uint8_t peek_byte( const std::size_t offset = 0 ) const noexcept
          {
             return static_cast< std::uint8_t >( peek_char( offset ) );