+++ include/tao/pegtl/internal/peek_mask_uint8.hpp
@@ -24,9 +24,9 @@ namespace tao
             using pair_t = input_pair< std::uint8_t >;
 
             template< typename Input >
-            static pair_t peek( Input& in, const std::size_t o = 0 ) noexcept( noexcept( in.peek_byte( 0 ) ) )
+            static pair_t peek( Input& in, const std::size_t o = 0 ) noexcept( noexcept( in.peek_uint8( 0 ) ) )
             {
-               return { std::uint8_t( in.peek_byte( o ) & M ), 1 };
+               return { std::uint8_t( in.peek_uint8( o ) & M ), 1 };
             }
          };