+++ include/tao/pegtl/memory_input.hpp
@@ -291,6 +291,12 @@ namespace tao
             return this->current()[ offset ];
          }
 
+         std::uint8_t peek_uint8( const std::size_t offset = 0 ) const noexcept
+         {
+            return static_cast< std::uint8_t >( peek_char( offset ) );
+         }
+
+         // Compatibility, remove with 3.0
          std::uint8_t peek_byte( const std::size_t offset = 0 ) const noexcept
          {
             return static_cast< std::uint8_t >( peek_char( offset ) );